# server/secure_comm_service/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Документация о внутренней реализации и логики работы находится в папке docs",
        "title": "SecureComm API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "paths": {
        "/files/all": {
            "get": {
                "security": [
                    {
                        "bearerAuth": []
                    }
                ],
                "description": "Возвращает пре‐подписанные ссылки на скачивание всех файлов заданной категории (photo, unknown, video, text).",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Files"
                ],
                "summary": "Получение всех файлов категории",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Bearer {token}",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Категория файлов (photo, unknown, video, text)",
                        "name": "type",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Список ссылок на все файлы категории",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/dto.FileResponse"
                            }
                        }
                    },
                    "400": {
                        "description": "Некорректная категория",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "403": {
                        "description": "Доступ запрещён",
                        "schema": {
                            "$ref": "#/definitions/cloud_handler.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Файлы не найдены",
                        "schema": {
                            "$ref": "#/definitions/cloud_handler.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Внутренняя ошибка сервера",
                        "schema": {
                            "$ref": "#/definitions/cloud_handler.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/files/many": {
            "delete": {
                "security": [
                    {
                        "bearerAuth": []
                    }
                ],
                "description": "Удаляет несколько объектов из MinIO, переданных в JSON-массиве, и снижает использование квоты.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Files"
                ],
                "summary": "Удаление нескольких файлов",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Bearer {token}",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "Массив идентификаторов объектов",
                        "name": "objectIDs",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.ObjectIDs"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Файлы успешно удалены",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "400": {
                        "description": "Некорректный JSON в теле запроса",
                        "schema": {
                            "$ref": "#/definitions/cloud_handler.ErrorResponse"
                        }
                    },
                    "403": {
                        "description": "Доступ запрещён",
                        "schema": {
                            "$ref": "#/definitions/cloud_handler.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Один из файлов не найден",
                        "schema": {
                            "$ref": "#/definitions/cloud_handler.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Внутренняя ошибка сервера",
                        "schema": {
                            "$ref": "#/definitions/cloud_handler.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/files/one": {
            "get": {
                "security": [
                    {
                        "bearerAuth": []
                    }
                ],
                "description": "Возвращает пре‐подписанную ссылку на скачивание одного файла по ID и типу.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Files"
                ],
                "summary": "Получение одного файла",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Bearer {token}",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Идентификатор объекта",
                        "name": "id",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Категория файла (photo, unknown, video, text)",
                        "name": "type",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Ссылка на скачивание файла",
                        "schema": {
                            "$ref": "#/definitions/dto.FileResponse"
                        }
                    },
                    "400": {
                        "description": "Некорректный запрос",
                        "schema": {
                            "$ref": "#/definitions/cloud_handler.ErrorResponse"
                        }
                    },
                    "403": {
                        "description": "Доступ запрещён",
                        "schema": {
                            "$ref": "#/definitions/cloud_handler.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Файл не найден",
                        "schema": {
                            "$ref": "#/definitions/cloud_handler.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Внутренняя ошибка сервера",
                        "schema": {
                            "$ref": "#/definitions/cloud_handler.ErrorResponse"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "bearerAuth": []
                    }
                ],
                "description": "Удаляет один объект из MinIO и снижает использование квоты.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Files"
                ],
                "summary": "Удаление одного файла",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Bearer {token}",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Идентификатор объекта",
                        "name": "id",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Категория файла (photo, unknown, video, text)",
                        "name": "type",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Файл успешно удалён",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "400": {
                        "description": "Некорректный запрос",
                        "schema": {
                            "$ref": "#/definitions/cloud_handler.ErrorResponse"
                        }
                    },
                    "403": {
                        "description": "Доступ запрещён",
                        "schema": {
                            "$ref": "#/definitions/cloud_handler.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Файл не найден",
                        "schema": {
                            "$ref": "#/definitions/cloud_handler.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Внутренняя ошибка сервера",
                        "schema": {
                            "$ref": "#/definitions/cloud_handler.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/files/one/encrypted": {
            "post": {
                "security": [
                    {
                        "bearerAuth": []
                    }
                ],
                "description": "Токен авторизации + зашифрованный поток в body + метаданные в заголовках",
                "consumes": [
                    "*/*"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Files"
                ],
                "summary": "Загрузка зашифрованного файла “на лету”",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Bearer {token}",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Оригинальное имя файла (например photo.jpg)",
                        "name": "X-Orig-Filename",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Оригинальный MIME-тип (например image/jpeg)",
                        "name": "X-Orig-Mime",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Категория файла (photo, video, text, unknown)",
                        "name": "X-File-Category",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Успешно: JSON с metadata + presigned URL",
                        "schema": {
                            "$ref": "#/definitions/dto.FileResponse"
                        }
                    },
                    "400": {
                        "description": "Некорректные заголовки или body",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "401": {
                        "description": "Проблема с авторизацией",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Внутренняя ошибка сервера",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "cloud_handler.ErrorResponse": {
            "type": "object",
            "properties": {
                "details": {
                    "description": "Дополнительные детали (при наличии)\nexample: \"field 'file' is required\""
                },
                "error": {
                    "description": "Краткое описание ошибки\nexample: \"Invalid request\"",
                    "type": "string"
                },
                "status": {
                    "description": "Код HTTP-статуса\nexample: 400",
                    "type": "integer"
                }
            }
        },
...
        "dto.FileResponse": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "mime_type": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "obj_id": {
                    "type": "string"
                },
                "url": {
                    "type": "string"
                }
            }
        },
...
        "dto.ObjectIDs": {
            "type": "object",
            "properties": {
                "object_ids": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "\"Bearer {token}\"",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}